+++ jinete/algorithms/heuristics/local_search/strategies/__init__.py
@@ -2,6 +2,6 @@ from .abc import (
     LocalSearchStrategy,
 )
 
-from .flips import (
+from .one_shift import (
     OneShiftLocalSearchStrategy,
 )
